When using method filtering, it turns out that if you provide an invalid method signature that results in CVM_TYPEID_ERROR, there is no warning message, and the entire class is accepted by the filtering mechanism. It is as if you just left off the method part and just listed the class name. This is because when we want to allow an entire class we stuff CVM_TYPEID_ERROR in the methodID field, and check for this in the filtering code.
My fix is to check for CVM_TYPEID_ERROR when looking up the method ID, and if there is one, force the classID to also be CVM_TYPEID_ERROR. I also print a warning message. I also added a check for a failed classID lookup, although I don't think that is possible as a result of an invalid class name.