Merged from components/cdc/branches/cdc-mr2-cr-5080490 using svn merge -r1267:HEAD https://phoneme.dev.java.net/svn/phoneme/components/cdc/branches/cdc-mr2-cr-5080490/
Continuation of fix for CR#5080490:
Fixed CVMmemCopy64() parameters to be CVMUInt32 * explicitly.
Also changed CCM runtime helpers to cast parameters to
CVMUint32 * instead of void * before calling CVMmemCopy64().
This fixes a build failure on some ports due to changes in the
last putback for this CR fix.  Tested with simulated builds
where CVMmemCopy64() is implemented with 2 32-bit copies.
Reviewed by Chris.